Resolve generics warnings in tests
There are several compiler warnings over the fact that
SchemaRegistryTransfer uses a generic type variable that is not
qualified in the test cases.

This commit replaces uses of the abstract `ConnectRecord` class with the
concrete `SourceRecord` class with no loss of correctness and enables
the missing Generics bindings to be set.